Added google credentials thrid party authentication, working flash messages within medicine.html and organSystems.html

+++ application.py
@@ -1,34 +1,24 @@
-from flask import Flask, render_template, url_for, redirect, request, jsonify
-from sqlalchemy import create_engine
+from flask import Flask, render_template, url_for, redirect, request, jsonify, flash
+from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Organ, Medicine
+# IMPORTS FOR GOOGLE credentials
+from flask import session as login_session
+import random
+import string
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
 
 app = Flask(__name__)
 
-##############################################################
-# This is only for iterative step #3 Templates and Forms
+CLIENT_ID = json.loads(
+    open('client_secret.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "RoadMapToHealth"
 
-# Adding Fake database using dictionary#################################
-'''
-# Fake organs
-organSystem = {'name': 'Digestive', 'id': '1'}
-
-organSystems = [{'name': 'Digestive', 'id': '1'}, {
-    'name': 'Heart', 'id': '2'}, {'name': 'Endocrine', 'id': '3'}]
-
-
-# Fake organs
-organ = {'name': 'The CRUDdy Crab', 'id': '1'}
-
-organs = [{'name': 'The CRUDdy Crab', 'id': '1'}, {
-    'name': 'Blue Burgers', 'id': '2'}, {'name': 'Taco Hut', 'id': '3'}]
-# Fake Menu Items
-items = [{'name': 'Cheese Pizza', 'description': 'made with fresh cheese', 'price': '$5.99', 'course': 'Entree', 'id': '1'}, {'name': 'Chocolate Cake', 'description': 'made with Dutch Chocolate', 'price': '$3.99', 'course': 'Dessert', 'id': '2'}, {'name': 'Caesar Salad', 'description': 'with fresh organic vegetables',
-                                                                                                                                                                                                                                                        'price': '$5.99', 'course': 'Entree', 'id': '3'}, {'name': 'Iced Tea', 'description': 'with lemon', 'price': '$.99', 'course': 'Beverage', 'id': '4'}, {'name': 'Spinach Dip', 'description': 'creamy dip with fresh spinach', 'price': '$1.99', 'course': 'Appetizer', 'id': '5'}]
-
-item = {'name': 'Cheese Pizza', 'description': 'made with fresh cheese',
-        'price': '$5.99', 'course': 'Entree'}
-'''
 
 ###########################################################################
 # Add our Database
@@ -39,6 +29,135 @@ Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+#########################################################################
+
+# Create anti-forgery state token
+
+
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    # return "The current session state is %s" % login_session['state']
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secret.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: '
+    print login_session['username']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
 #########################################################################
 # JSON API's
 
@@ -82,12 +201,12 @@ def showOrganSystems():
 @app.route('/RoadMapToHealth/new/', methods=['GET', 'POST'])
 def newOrganSystem():
     # protecting the pages
-    # if  'username' not in login_session:
-    #    return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         newOrgan = Organ(name=request.form['name'])
         session.add(newOrgan)
-        #flash('New organ %s Successfully Created' % neworgan.name)
+        flash('New organ %s Successfully Created' % neworgan.name)
         session.commit()
         return redirect(url_for('showOrganSystems'))
     else:
@@ -98,12 +217,15 @@ def newOrganSystem():
 @app.route('/RoadMapToHealth/<int:organ_id>/edit/', methods=['GET', 'POST'])
 def editOrganSystem(organ_id):  # organSystem_id
     # return render_template('editOrgan.html', organSystem=organSystem)
+    # page protection
+    if 'username' not in login_session:
+        return redirect('/login')
     editedOrgan = session.query(
         Organ).filter_by(id=organ_id).one()
     if request.method == 'POST':
         if request.form['name']:
             editedOrgan.name = request.form['name']
-            #flash('Organ Successfully Edited %s' % editedOrgan.name)
+            flash('Organ %s successfully edited ' % editedOrgan.name)
             return redirect(url_for('showOrganSystems'))
     else:
         return render_template('editOrgan.html', organ=editedOrgan, organ_id=organ_id)
@@ -115,11 +237,14 @@ def editOrganSystem(organ_id):  # organSystem_id
 @app.route('/RoadMapToHealth/<int:organ_id>/delete/', methods=['GET', 'POST'])
 def deleteOrganSystem(organ_id):  # OrganSystem_id
     # return render_template('deleteOrgan.html', organSystem=organSystem)
+    # protecting the page
+    if 'username' not in login_session:
+        return redirect('/login')
     organToDelete = session.query(
         Organ).filter_by(id=organ_id).one()
     if request.method == 'POST':
         session.delete(organToDelete)
-        #flash('%s Successfully Deleted' % organToDelete.name)
+        flash('%s Successfully Deleted' % organToDelete.name)
         session.commit()
         return redirect(url_for('showOrganSystems', organ_id=organ_id))
     else:
@@ -143,13 +268,16 @@ def showMedicine(organ_id):  # organSystem_id
 @app.route('/RoadMapToHealth/<int:organ_id>/medicine/new/', methods=['GET', 'POST'])
 def newMedicine(organ_id):
     # return render_template('newMedicine.html')
+    # Page Protection
+    if 'username' not in login_session:
+        return redirect('/login')
     organ = session.query(Organ).filter_by(id=organ_id).one()
     if request.method == 'POST':
         newItem = Medicine(name=request.form['name'], description=request.form[
                            'description'], type=request.form['type'], gland=request.form['gland'], organ_id=organ_id, organ=organ)
         session.add(newItem)
         session.commit()
-        #flash('New Menu %s Item Successfully Created' % (newItem.name))
+        flash('New Medicine %s Successfully Created' % (newItem.name))
         return redirect(url_for('showMedicine', organ_id=organ_id))
     else:
         return render_template('newMedicine.html', organ_id=organ_id)
@@ -160,6 +288,9 @@ def newMedicine(organ_id):
 @app.route('/RoadMapToHealth/<int:organ_id>/medicine/<int:medicine_id>/edit', methods=['GET', 'POST'])
 def editMedicine(organ_id, medicine_id):  # organSystem_id,medicine_id
     # return render_template('editMedicine.html', item=item)
+    # Page Protection
+    if 'username' not in login_session:
+        return redirect('/login')
     editedItem = session.query(Medicine).filter_by(id=medicine_id).one()
     organ = session.query(Organ).filter_by(id=organ_id).one()
     if request.method == 'POST':
@@ -173,7 +304,7 @@ def editMedicine(organ_id, medicine_id):  # organSystem_id,medicine_id
             editedItem.course = request.form['type']
         session.add(editedItem)
         session.commit()
-        #flash('Menu Item Successfully Edited')
+        flash('Medicine %s Successfully Edited' % (editedItem.name))
         return redirect(url_for('showMedicine', organ_id=organ_id))
     else:
         return render_template('editMedicine.html', organ_id=organ_id, medicine_id=medicine_id, item=editedItem)
@@ -185,17 +316,21 @@ def editMedicine(organ_id, medicine_id):  # organSystem_id,medicine_id
 @app.route('/RoadMapToHealth/<int:organ_id>/medicine/<int:medicine_id>/delete', methods=['GET', 'POST'])
 def deleteMedicine(organ_id, medicine_id):  # organSystem_id,medicine_id
     # return render_template('deleteMedicine.html', item=item)
+    # page Protection
+    if 'username' not in login_session:
+        return redirect('/login')
     organ = session.query(Organ).filter_by(id=organ_id).one()
     itemToDelete = session.query(Medicine).filter_by(id=medicine_id).one()
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
-    #        flash('Menu Item Successfully Deleted')
+        flash('Medicine Successfully Deleted')
         return redirect(url_for('showMedicine', organ_id=organ_id))
     else:
         return render_template('deleteMedicine.html', item=itemToDelete, organ_id=organ_id)
 
 
 if __name__ == '__main__':
+    app.secret_key = 'super_secret_key'
     app.debug = True
     app.run(host='0.0.0.0', port=8000)
